Refactored some UI rendering code and large Event Handling changes
"Exercise buttons" are changed to "exercise cards", which consists of a div container, a button, a label, and an input.

The label and input are hidden by default.

When a button is clicked, its sibling label and input is unhidden

When a different button is clicked, the previously clicked buttons siblings are re-hidden and the currently clicked button's siblings are shown