Updated code to handle expanding/closing of section in the background.
When a user creates a new Exercise, the routine section now opens once it's saved. This is to help to help the user see where their Exercise can be interacted with.

When a user creates a new Routine, the exercise AND routine section closes so that the user can focus on the timer.

They can manually expand/close the sections at any time.